requirements.txt
altair==4.2.2
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
audioread==3.0.0
beautifulsoup4==4.11.2
blinker==1.5
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
decorator==4.4.2
Django==3.2.17
entrypoints==0.4
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
librosa==0.9.2
llvmlite==0.39.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
moviepy==1.0.3
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
proglog==0.1.10
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytube3==9.6.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
resampy==0.4.2
rich==13.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
soundfile==0.11.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
streamlit==1.18.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
whitenoise==6.3.0
zipp==3.13.0
